Renaming all Android references
