+++ android/app/src/main/java/com/magicbell/mobileinbox/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.rnprototype
+package com.magicbell.mobileinbox
 import expo.modules.ReactActivityDelegateWrapper
 
 import com.facebook.react.ReactActivity
@@ -12,7 +12,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "RNProtoType"
+  override fun getMainComponentName(): String = "MobileInbox"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
