+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">RNProtoType</string>
+    <string name="app_name">Mobile Inbox</string>
 </resources>
